% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_functions.R
\name{check_PyH_setup}
\alias{check_PyH_setup}
\title{Make sure that python has been configured correctly for persistent homology calculations.}
\usage{
check_PyH_setup()
}
\description{
Ensures that the reticulate package, python and the python module "ripser" have all been installed.
}
\author{
Shael Brown - \email{shaelebrown@gmail.com}
}
