% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_functions.R
\name{check_ripser}
\alias{check_ripser}
\title{Verify an imported ripser module.}
\usage{
check_ripser(ripser)
}
\arguments{
\item{ripser}{the ripser module object.}
}
\description{
Verify an imported ripser module.
}
\author{
Shael Brown - \email{shaelebrown@gmail.com}
}
