% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{D1}
\alias{D1}
\title{A small persistence diagram.}
\format{
A data frame with 1 row and 3 variables:
\describe{
  \item{dimension}{homological dimension of the topological feature}
  \item{birth}{birth radius of the topological feature}
  \item{death}{death radius of the topological feature}
}
}
\usage{
D1
}
\description{
A dataset containing the dimension, birth and death values for
a small persistence diagram.
}
\keyword{datasets}
