% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{D2}
\alias{D2}
\title{A small persistence diagram.}
\format{
A data frame with 2 rows and 3 variables:
\describe{
  \item{dimension}{homological dimension of the topological features}
  \item{birth}{birth radius of the topological features}
  \item{death}{death radius of the topological features}
}
}
\usage{
D2
}
\description{
A dataset containing the dimension, birth and death values for
a small persistence diagram.
}
\keyword{datasets}
