% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_calculations.R
\name{diagram_kernel}
\alias{diagram_kernel}
\title{Calculate persistence Fisher kernel between a pair of persistence diagrams}
\usage{
diagram_kernel(dim = 0, sigma = 1, t = 1)
}
\arguments{
\item{dim}{the homological dimension in which the distance is to be computed.}

\item{sigma}{a positive number representing the bandwith for the Fisher information metric, default 1.}

\item{t}{a positive number representing the scale for the kernel, default 1.}
}
\value{
the kernel function.
}
\description{
Returns a function which calculates the persistence Fisher kenel between a pair of persistce diagrams, stored as
data frames (as the output from diagram_to_df) in a particular homological dimension.
}
\details{
The `dim` parameter should be a positive finite integer.
The `sigma` parameter is the positive bandwith for the Fisher information metric, and
`t` is the scale parameter for persistence Fisher kernel.
}
\examples{

# create two diagrams with package TDA based on 2D Gaussians
diag1 <- TDA::ripsDiag(data.frame(x = rnorm(100,mean = 0,sd = 1),
y = rnorm(100,mean = 0,sd = 1)),
maxscale = 1,
maxdimension = 1)
diag2 <- TDA::ripsDiag(data.frame(x = rnorm(100,mean = 0,sd = 1),
y = rnorm(100,mean = 0,sd = 1)),
maxscale = 1,
maxdimension = 1)

# calculate their kernel value in dimension 1 with sigma = 2, t = 2
fisher_kernel <- diagram_kernel(dim = 1,sigma = 2,t = 2)
k <- fisher_kernel(D1 = diag1,D2 = diag2)
}
