% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{diagram_kkmeans}
\alias{diagram_kkmeans}
\title{Cluster a group of persistence diagrams using kernel k-means}
\usage{
diagram_kkmeans(
  diagrams,
  centers,
  dim = 0,
  t = 1,
  sigma = 1,
  num_workers = parallelly::availableCores(omit = 1),
  ...
)
}
\arguments{
\item{diagrams}{a list of persistence diagrams, as the output of a TDA calculation.}

\item{centers}{number of clusters to initialize.}

\item{dim}{the homological dimension in which the distance is to be computed.}

\item{t}{the positive scale for the persistence Fisher kernel, default 1.}

\item{sigma}{a positive number representing the bandwith for the Fisher information metric, default 1}

\item{num_workers}{the number of cores used for parallel computation, default is one less the number of cores on the machine.}

\item{...}{additional parameters.}
}
\value{
a 'diagram_kkmeans' object containing the output of \code{\link[kernlab]{kkmeans}} on the diagram distance matrix, i.e. a list containing

\describe{

\item{clustering}{an S4 object of class specc, the output of a \code{\link[kernlab]{kkmeans}} function call. The .Data slot of this object contains cluster memberships, withinss contains the within-cluster sum of squares for each cluster, etc.}

\item{diagrams}{the input `diagrams` argument.}

\item{dim}{the input `dim` argument.}

\item{t}{the input `t` argument.}

\item{sigma}{the input `sigma` argument.}

}
}
\description{
Returns the output of kkmeans on the desired Gram matrix of a group of persistence diagrams
in a particular dimension.
}
\details{
The `diagrams` parameter should be a list of persistence diagrams computed from TDA.
The `dim` parameter should be a positive finite integer.
The `sigma` and `t` parameters are the positive bandwith for the Fisher information metric and
the positive scale for the persistence Fisher kernel respectively. `num_workers` is the
number of cores used for parallel computation.
`centers` is the number of desired clusters, and
`...` are additional parameters to the kkmeans kernlab function.
}
\examples{

# create nine diagrams from three base diagrams
g <- generate_TDAML_test_data(3,3,3)

# calculate kmeans clusters with centers = 3, and sigma = t = 2
clust <- diagram_kkmeans(diagrams = g,centers = 3,dim = 0,t = 2,sigma = 2,num_workers = 2)
}
\seealso{
\code{\link{diagram_nearest_clusters}} for predicting clusters of new diagrams.
}
\author{
Shael Brown - \email{shaelebrown@gmail.com}
}
