% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{diagram_MDS}
\alias{diagram_MDS}
\title{Calculate the metric multidimensional scaling of a group of persistence diagrams}
\usage{
diagram_MDS(
  diagrams,
  distance = "wasserstein",
  dim = 0,
  p = 2,
  sigma = NULL,
  k = 2,
  eig = FALSE,
  add = FALSE,
  x.ret = FALSE,
  list. = eig || add || x.ret,
  num_workers = parallelly::availableCores(omit = 1)
)
}
\arguments{
\item{diagrams}{a list of persistence diagrams, as the output of a TDA calculation like \code{\link[TDA]{ripsDiag}} or a \code{\link{diagram_to_df}} function call.}

\item{distance}{a string representing the desired distance metric to be used, either 'wasserstein' (default) or 'fisher'.}

\item{dim}{the homological dimension in which the distance is to be computed.}

\item{p}{the wasserstein power, a number at least 1 (infinity for the bottleneck distance), default 2.}

\item{sigma}{a positive number representing the bandwith for the Fisher information metric, default NULL.}

\item{k}{the maximum dimension of the space which the data are to be represented in; must be in {1,2,...,n-1}.}

\item{eig}{indicates whether eigenvalues should be returned.}

\item{add}{logical indicating if an additive constant c* should be computed, and added to the non-diagonal dissimilarities such that the modified dissimilarities are Euclidean.}

\item{x.ret}{indicates whether the doubly centered symmetric distance matrix should be returned.}

\item{list.}{local indicating if a list should be returned or just the n*k matrix.}

\item{num_workers}{the number of cores used for parallel computation, default is one less the number of cores on the machine.}
}
\value{
the output of \code{\link[stats]{cmdscale}} on the diagram distance matrix. If `list.` is false (as per default),
a matrix with `k` columns whose rows give the coordinates of the points chosen to represent the dissimilarities.

Otherwise, a list containing the following components.

\describe{

 \item{points}{a matrix with up to `k` columns whose rows give the coordinates of the points chosen to represent the dissimilarities.}

 \item{eig}{the \eqn{n} eigenvalues computed during the scaling process if `eig` is true.}
 
 \item{x}{the doubly centered distance matrix if `x.ret` is true.}
 
 \item{ac}{the additive constant \eqn{c^*}, 0 if `add` = FALSE.}
 
 \item{GOF}{the numeric vector of length 2, representing the sum of all the eigenvalues divided by the sum of their absolute values (first vector element) or by the sum of the max of each eigenvalue and 0 (second vector element).}

}
}
\description{
Returns the output of cmdscale on the desired distance matrix of a group of persistence diagrams
in a particular dimension.
}
\details{
The `diagrams` parameter should be a list of persistence diagrams computed from a TDA calculation like \code{\link[TDA]{ripsDiag}} or from \code{\link{diagram_to_df}}.
The `distance` parameter is a string representing which determines which distance metric to use.
The `dim` parameter should be a positive finite integer.
The `sigma` parameter is the positive bandwith for the Fisher information metric, and
`p` is the wasserstein power parameter. `k` is the desired dimension of the embedding, and
`eig`, `add`, `x.ret` and `list.` are cmdscale parameters providing optional additional
information to be returned. `num_workers` is the number of cores used for parallel computation.
}
\examples{

# create nine diagrams from three base diagrams
g <- generate_TDAML_test_data(3,3,3)

# calculate their 2D MDS embedding in dimension 1 with the bottleneck distance
mds <- diagram_MDS(diagrams = g,dim = 0,p = Inf,k = 2,num_workers = 2)
}
\author{
Shael Brown - \email{shaelebrown@gmail.com}
}
