% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generate_TDAML_test_data}
\alias{generate_TDAML_test_data}
\title{Creates persistence diagrams to test TDAML functions.}
\usage{
generate_TDAML_test_data(num_D1, num_D2, num_D3)
}
\arguments{
\item{num_D1}{the number of desired noisy copies of D1.}

\item{num_D2}{the number of desired noisy copies of D2.}

\item{num_D3}{the number of desired noisy copies of D3.}
}
\value{
either a single diagram, or a list of `num_D1` copies of D1, 
`num_D2` copies of D2 and `num_D3` copies of D3
with independent Gaussian noise added to the birth and death values, 
each with variance 0.05^2.
}
\description{
An internal function which uses the three example persistence diagrams,
each with points only in dimension 0, to create a list of diagrams for 
further analysis. The first diagram (D1) has one point with birth value
2 and death value 3. The second diagram (D2) has two points, the first
has birth value 2 and death value 3.3, and the second has birth value
0 and death value 0.5. The third diagram (D3) has one point with birth
value 0 and death value 0.5. The function either returns one of these
three diagrams, or a list of diagrams which were each generated by one
of D1, D2 and D3 but having Gaussian noise added to the birth and death
values for each point independently (with variance 0.05^2).
}
\details{
The `num_D1`, `num_D2` and `num_D3` parameters should be the number of desired copies
of D1, D2 and D3 respectively. If exactly one of these parameters is 1 and the others
are 0 then just return the noiseless diagram corresponding to that parameter.
}
\examples{

# generate just D1
D1 <- generate_TDAML_test_data(1,0,0)

# create three copies of each of D1, D2 and D3
l <- generate_TDAML_test_data(3,3,3)
}
\references{

}
\author{
Shael Brown - \email{shaelebrown@gmail.com}
}
\keyword{internal}
