% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_functions.R
\name{import_ripser}
\alias{import_ripser}
\title{Import the python module ripser for fast persistent homology calculations.}
\usage{
import_ripser()
}
\description{
Same as "reticulate::import("ripser")".
}
\examples{
# set up python for fast PH
PyH_setup()

# import ripser
ripser <- import_ripser()
}
\author{
Shael Brown - \email{shaelebrown@gmail.com}
}
