% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{noisy_copies_D1_D2_D3}
\alias{noisy_copies_D1_D2_D3}
\title{Creates persistence diagrams from D1, D2 and D3 with random (Gaussian) noise
added to each point in each diagram.}
\usage{
noisy_copies_D1_D2_D3(num_D1, num_D2, num_D3)
}
\arguments{
\item{num_D1}{the number of desired noisy copies of D1.}

\item{num_D2}{the number of desired noisy copies of D2.}

\item{num_D3}{the number of desired noisy copies of D3.}
}
\value{
a list of `num_D1` copies of D1, `num_D2` copies of D2 and `num_D3` copies of D3
with independent Gaussian noise added to the birth and death values, each with variance 0.05^2.
}
\description{
Uses the three example persistence diagrams to create a list of diagrams
for further analysis. Intended for use within this package, however can also
be used for replicating results.
}
\details{
The `num_D1`, `num_D2` and `num_D3` parameters should be the number of desired copies
of D1, D2 and D3 respectively.
}
\examples{

# create three copies of each of D1, D2 and D3
noisy_copies_D1_D2_D3(3,3,3)
}
\references{

}
\author{
Shael Brown - \email{shaelebrown@gmail.com}
}
