% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rips_complexes.R
\name{plot_rips_graph}
\alias{plot_rips_graph}
\title{Plot a Rips graph using the \link{igraph} package.}
\usage{
plot_rips_graph(graphs, index = 1L)
}
\arguments{
\item{graphs}{the output of a `rips_graphs` function call.}

\item{index}{the integer representing which graph out of the `graphs` parameter should be plotted, default 1.}
}
\description{
This function will throw an error if the \link{igraph} package is not installed.
}
\examples{

if(require("TDA") & require("igraph"))
{
  # simulate data from the unit circle and calculate 
  # its diagram
  df <- TDA::circleUnif(n = 25)
  diag <- TDA::ripsDiag(df,maxdimension = 1,maxscale = 2)
  
  # get minimum death radius of any data cluster
  min_death_H0 <- 
  min(diag$diagram[which(diag$diagram[,1] == 0),3L])
  
  # get birth and death radius of the loop
  loop_birth <- as.numeric(diag$diagram[nrow(diag$diagram),2L])
  loop_death <- as.numeric(diag$diagram[nrow(diag$diagram),3L])

  # compute Rips-Vietoris graphs at radii half of 
  # min_death_H0 and the mean of loop_birth and 
  # loop_death, returning clusters
  graphs <- rips_graphs(X = df,eps = 
  c(0.5*min_death_H0,(loop_birth + loop_death)/2))
  
  # plot graph of smaller (first) radius
  plot_rips_graph(graphs = graphs,index = 1L)
  
  # plot graph of larger (second) radius
  plot_rips_graph(graphs = graphs,index = 2L)
  
  # repeat but with rownames for df, each vertex
  # will be plotted with its rownames
  rownames(df) <- paste0("V",1:25)
  graphs <- rips_graphs(X = df,eps = 
  c(0.5*min_death_H0,(loop_birth + loop_death)/2))
  plot_rips_graph(graphs = graphs,index = 1L)
  
}
}
\author{
Shael Brown - \email{shaelebrown@gmail.com}
}
