% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_functions.R
\name{PyH_setup}
\alias{PyH_setup}
\title{Configure R to interface with python for fast persistent homology calculations.}
\usage{
PyH_setup()
}
\description{
Installs the "reticulate" package, python and the python module "ripser".
}
\details{
The "ripser" python module has faster persistent homology calculations than the R wrapper for ripser, the TDAstats
package, even when using the reticulate package as an interface. This function allows
access to the added python speedup, and only needs to be called one time after TDApplied has been
installed (not at the beginning of every R session).
}
\examples{

# set up python for fast PH
PyH_setup()
}
\author{
Shael Brown - \email{shaelebrown@gmail.com}
}
