% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{TDA_diagram_to_df}
\alias{TDA_diagram_to_df}
\title{Convert TDA persistence diagrams to data frames}
\usage{
TDA_diagram_to_df(d)
}
\arguments{
\item{d}{output of a homology calculation, like ripsDiag}
}
\value{
3-column data frame, with each row representing a topological feature
}
\description{
The output of homology calculations in package TDA
are not immediately able to be used for distance calculations.
These functions convert the output of homology calculations in TDA
into a data frame for further usage.
}
\details{
The `d` parameter is the persistence diagram to be converted.
}
\examples{

# create a persistence diagram from a 2D Gaussian using TDA
df = data.frame(x = rnorm(n = 100,mean = 0,sd = 1),y = rnorm(n = 100,mean = 0,sd = 1))
# compute persistence barcode with ripsDiag
phom = ripsDiag(X = df,maxdimension = 1,maxscale = 1)
# convert to data frame
phom_df = TDA_diagram_to_df(d = phom)

# create a persistence diagram from a 2D Gaussian using TDAstats
df = data.frame(x = rnorm(n = 100,mean = 0,sd = 1),y = rnorm(n = 100,mean = 0,sd = 1))

# compute persistence barcode with calculate_homology
phom = calculate_homology(mat = df,maxdimension = 1,maxscale = 1)

# convert to data frame
phom_df = TDAstats_diagram_to_df(d = phom,dim = 1,threshold = 1)
}
