% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDAInference.R
\docType{package}
\name{TDAInference}
\alias{TDAInference}
\title{Statistical Inference for Persistence Diagrams}
\description{
This package provides the functionality to infer when
there are topological differences between an arbitrary
number of groups of persistence diagrams, and to quickly
compute distances between persistence diagrams.
}
