% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDApplied-package.R
\docType{package}
\name{TDApplied-package}
\alias{TDApplied}
\alias{TDApplied-package}
\title{TDApplied: Machine Learning and Inference for Topological Data Analysis}
\description{
Topological data analysis is a powerful tool for finding non-linear global structure in whole datasets. The main tool of topological data analysis is persistent homology, which computes a topological shape descriptor of a dataset called a persistence diagram. 'TDApplied' provides useful and efficient methods for analyzing groups of persistence diagrams with machine learning and statistical inference, and these functions can also interface with other data science packages to form flexible and integrated topological data analysis pipelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shaelebrown/TDApplied}
  \item Report bugs at \url{https://github.com/shaelebrown/TDApplied/issues}
}

}
\author{
\strong{Maintainer}: Shael Brown \email{shaelebrown@gmail.com}

Authors:
\itemize{
  \item Dr. Reza Farivar \email{reza.farivar@mcgill.ca} [funder]
}

}
\keyword{internal}
