% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{TDAstats_diagram_to_df}
\alias{TDAstats_diagram_to_df}
\title{Convert TDAstats persistence diagrams to data frames}
\usage{
TDAstats_diagram_to_df(d)
}
\arguments{
\item{d}{output of a calculate_homology calculation}
}
\value{
3-column data frame, with each row representing a topological feature
}
\description{
The output of homology calculations in package TDAstats
are not immediately able to be used for distance calculations.
These functions convert the output of homology calculations in TDAstats
into a data frame for further usage.
}
\details{
The `d` parameter is the persistence diagram to be converted.
}
\examples{

# create a persistence diagram from a 2D Gaussian using TDAstats
df = data.frame(x = rnorm(n = 100,mean = 0,sd = 1),y = rnorm(n = 100,mean = 0,sd = 1))

# compute persistence barcode with calculate_homology
phom = calculate_homology(mat = df,maxdimension = 1,maxscale = 1)

# convert to data frame
phom_df = TDAstats_diagram_to_df(d = phom,dim = 1,threshold = 1)
}
