% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_calculations.R
\name{diagram_kernel}
\alias{diagram_kernel}
\title{Calculate persistence Fisher kernel between a pair of persistence diagrams}
\usage{
diagram_kernel(D1, D2, dim = 0, sigma = 1, t = 1)
}
\arguments{
\item{D1}{the first persistence diagram, either outputted from a TDA calculation like \code{\link[TDA]{ripsDiag}} or from a \code{\link{diagram_to_df}} function call.}

\item{D2}{the second persistence diagram, either outputted from a TDA calculation like \code{\link[TDA]{ripsDiag}} or from a \code{\link{diagram_to_df}} function call.}

\item{dim}{the homological dimension in which the distance is to be computed.}

\item{sigma}{a positive number representing the bandwith for the Fisher information metric, default 1.}

\item{t}{a positive number representing the scale for the persistence Fisher kernel, default 1.}
}
\value{
the numeric kernel value.
}
\description{
Returns the persistence Fisher kenel between a pair of persistence diagrams
in a particular homological dimension.
}
\details{
The `D1` and `D2` parameters are the two persistence diagrams.
The `dim` parameter should be a positive finite integer.
The `sigma` parameter is the positive bandwith for the Fisher information metric, and
`t` is the scale parameter for the persistence Fisher kernel.
}
\examples{

# load three diagrams
D1 <- generate_TDAML_test_data(1,0,0)
D2 <- generate_TDAML_test_data(0,1,0)
D3 <- generate_TDAML_test_data(0,0,1)

# calculate the kernel value between D1 and D2 with sigma = 2, t = 2
diagram_kernel(D1,D2,dim = 0,sigma = 2,t = 2)
# calculate the kernel value between D1 and D3 with sigma = 2, t = 2
diagram_kernel(D1,D3,dim = 0,sigma = 2,t = 2)
}
\references{
Le T, Yamada M (2018). "Persistence fisher kernel: a riemannian manifold kernel for persistence diagrams." \url{https://proceedings.neurips.cc/paper/2018/file/959ab9a0695c467e7caf75431a872e5c-Paper.pdf}.
}
\author{
Shael Brown - \email{shaelebrown@gmail.com}
}
