% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{diagram_kkmeans}
\alias{diagram_kkmeans}
\title{Cluster a group of persistence diagrams using kernel k-means}
\usage{
diagram_kkmeans(diagrams, centers, dim = 0, t = 1, sigma = 1, ...)
}
\arguments{
\item{diagrams}{a list of persistence diagrams, as the output of a TDA calculation.}

\item{centers}{number of clusters to initialize.}

\item{dim}{the homological dimension in which the distance is to be computed.}

\item{t}{the positive scale for the persistence Fisher kernel, default 1.}

\item{sigma}{a positive number representing the bandwith for the Fisher information metric, default 1}

\item{...}{additional parameters.}
}
\value{
a diagram_kkmeans object containing the output of \code{\link[kernlab]{kkmeans}} on the diagram distance matrix, i.e. the cluster memberships, centers, sizes, and withinss's, and the diagrams, dim, t and sigma parameters.
}
\description{
Returns the output of kkmeans on the desired Gram matrix of a group of persistence diagrams
in a particular dimension.
}
\details{
The `diagrams` parameter should be a list of persistence diagrams computed from TDA.
The `dim` parameter should be a positive finite integer.
The `sigma` and `t` parameters are the positive bandwith for the Fisher information metric and
the positive scale for the persistence Fisher kernel respectively.
`centers` is the number of desired clusters, and
`...` are additional parameters to the kkmeans kernlab function.
}
\examples{

# create ten diagrams with package TDA based on 2D Gaussians
g <- lapply(X = 1:10,FUN = function(X){

diag <- TDA::ripsDiag(data.frame(x = rnorm(100,mean = 0,sd = 1),
y = rnorm(100,mean = 0,sd = 1)),
maxscale = 1,
maxdimension = 1)
df <- diagram_to_df(d = diag)
return(df)

})

# calculate kmeans clusters with centers = 2 in dimension 1 with sigma = t = 2
clusters <- diagram_kkmeans(diagrams = g,centers = 2,dim = 1,t = 2,sigma = 2)
}
