% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{diagram_MDS}
\alias{diagram_MDS}
\title{Calculate the metric multidimensional scaling of a group of persistence diagrams}
\usage{
diagram_MDS(
  diagrams,
  distance = "wasserstein",
  dim = 0,
  p = 2,
  sigma = NULL,
  k = 2,
  eig = FALSE,
  add = FALSE,
  x.ret = FALSE,
  list. = eig || add || x.ret
)
}
\arguments{
\item{diagrams}{a list of persistence diagrams, as the output of a TDA calculation like \code{\link[TDA]{ripsDiag}} or a \code{\link{diagram_to_df}} function call.}

\item{distance}{a string representing the desired distance metric to be used, either 'wasserstein' (default) or 'fisher'.}

\item{dim}{the homological dimension in which the distance is to be computed.}

\item{p}{the wasserstein power, a number at least 1 (infinity for the bottleneck distance), default 2.}

\item{sigma}{a positive number representing the bandwith for the Fisher information metric, default NULL.}

\item{k}{the maximum dimension of the space which the data are to be represented in; must be in {1,2,...,n-1}.}

\item{eig}{indicates whether eigenvalues should be returned.}

\item{add}{logical indicating if an additive constant c* should be computed, and added to the non-diagonal dissimilarities such that the modified dissimilarities are Euclidean.}

\item{x.ret}{indicates whether the doubly centered symmetric distance matrix should be returned.}

\item{list.}{local indicating if a list should be returned or just the n*k matrix.}
}
\value{
the output of \code{\link[stats]{cmdscale}} on the diagram distance matrix, either just the embedding matrix or a list containing the embedding and additional information.
}
\description{
Returns the output of cmdscale on the desired distance matrix of a group of persistence diagrams
in a particular dimension.
}
\details{
The `diagrams` parameter should be a list of persistence diagrams computed from a TDA calculation like \code{\link[TDA]{ripsDiag}} or from \code{\link{diagram_to_df}}.
The `distance` parameter is a string representing which determines which distance metric to use.
The `dim` parameter should be a positive finite integer.
The `sigma` parameter is the positive bandwith for the Fisher information metric, and
`p` is the wasserstein power parameter. `k` is the desired dimension of the embedding, and
`eig`, `add`, `x.ret` and `list.` are cmdscale parameters providing optional additional
information to be returned.
}
\examples{

# create ten diagrams with package TDA based on 2D Gaussians
g <- lapply(X = 1:10,FUN = function(X){

diag <- TDA::ripsDiag(data.frame(x = rnorm(100,mean = 0,sd = 1),
y = rnorm(100,mean = 0,sd = 1)),
maxscale = 1,
maxdimension = 1)
df <- diagram_to_df(d = diag)
return(df)

})

# calculate their 2D mds embedding in dimension 1 with the bottleneck distance
embedding <- diagram_MDS(diagrams = g,dim = 1,p = Inf,k = 2)
}
