% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_calculations.R
\name{gram_matrix}
\alias{gram_matrix}
\title{Calculate the Gram matrix K for a list of persistence diagrams, i.e. K[i,j] = k(d[i],d[j]), in parallel.}
\usage{
gram_matrix(diagrams, dim = 0, sigma = 1, t = 1)
}
\arguments{
\item{diagrams}{the list of persistence diagrams, either the output from TDA calculations or the diagram_to_df function.}

\item{dim}{the homological dimension in which the distance is to be computed.}

\item{sigma}{a positive number representing the bandwith for the Fisher information metric, default 1.}

\item{t}{a positive number representing the scale for the kernel, default 1.}
}
\value{
the Gram matrix of class 'kernelMatrix' for downstream analyses with the kernlab package.
}
\description{
Returns the Gram matrix for a list of persistence diagrams.
}
\details{
`diagrams` is the list of persistence diagrams.
The `dim` parameter should be a positive finite integer.
The `sigma` parameter is the positive bandwith for the Fisher information metric, and
`t` is the positive scale parameter for the persistence Fisher kernel.
}
\examples{

# create ten diagrams with package TDA based on 2D Gaussians
g <- lapply(X = 1:10,FUN = function(X){

diag <- TDA::ripsDiag(data.frame(x = rnorm(100,mean = 0,sd = 1),
y = rnorm(100,mean = 0,sd = 1)),
maxscale = 1,
maxdimension = 1)
df <- diagram_to_df(d = diag)
return(df)

})

# calculate their Gram matrix in dimension 1 with sigma = 2, t = 2
G <- gram_matrix(diagrams = g,dim = 1,sigma = 2,t = 2)
}
