% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{predict_diagram_kpca}
\alias{predict_diagram_kpca}
\title{Project persistence diagrams into a low-dimensional space via a pre-computed kernel PCA embedding.}
\usage{
predict_diagram_kpca(
  new_diagrams,
  embedding,
  num_workers = parallelly::availableCores(omit = 1)
)
}
\arguments{
\item{new_diagrams}{a list of persistence diagrams which are the output of a TDA calculation like \code{\link[TDA]{ripsDiag}} or \code{\link{diagram_to_df}}.}

\item{embedding}{the output of a \code{\link{diagram_kpca}} function call.}

\item{num_workers}{the number of cores used for parallel computation, default is one less than the number of cores on the machine.}
}
\value{
the data projection (rotation), stored as a numeric matrix. Each row corresponds to the same-index diagram in `new_diagrams`.
}
\description{
Compute the location in low-dimensional space of each element of a list of new persistence diagrams using a
previously-computed kernel PCA embedding (from the \code{\link{diagram_kpca}} function).
}
\examples{

# create nine diagrams from three base diagrams
g <- generate_TDAML_test_data(3,3,3)

# calculate their 2D PCA embedding with sigma = t = 2
pca <- diagram_kpca(diagrams = g,dim = 0,t = 2,sigma = 2,features = 2,num_workers = 2)

# project new diagrams onto old model
g_new <- generate_TDAML_test_data(3,3,3)

# calculate new embedding coordinates
new_pca <- predict_diagram_kpca(new_diagrams = g_new,embedding = pca,num_workers = 2)
}
\seealso{
\code{\link{diagram_kpca}} for embedding persistence diagrams into a low-dimensional space.
}
\author{
Shael Brown - \email{shaelebrown@gmail.com}
}
