% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine_learning.R
\name{predict_diagram_ksvm}
\alias{predict_diagram_ksvm}
\title{Predict the response to a new group of persistence diagrams from a computed diagram_ksvm model}
\usage{
predict_diagram_ksvm(new_diagrams, model)
}
\arguments{
\item{new_diagrams}{a list of new persistence diagrams, as the output of a TDA calculation.}

\item{model}{the diagram_ksvm model to be used for prediction.}
}
\value{
a vector containing the output of \code{\link[kernlab]{predict.ksvm}} on the cross Gram matrix of the new diagrams and the support vector diagrams stored in the model.
}
\description{
Returns the predicted response vector of the model on the new diagrams.
}
\details{
The `new_diagrams` parameter should be a list of persistence diagrams computed from a TDA calculation like \code{\link[TDA]{ripsDiag}} or \code{\link{diagram_to_df}}.
The `model` parameter should be the output from a diagram_ksvm function call.
}
\examples{

# create ten diagrams with package TDA based on 2D Gaussians
g <- lapply(X = 1:10,FUN = function(X){

diag <- TDA::ripsDiag(data.frame(x = rnorm(100,mean = 0,sd = 1),
y = rnorm(100,mean = 0,sd = 1)),
maxscale = 1,
maxdimension = 1)
df <- diagram_to_df(d = diag)
return(df)

})

# create random response vector
y <- stats::runif(10,min = 0,max = 1)

# calculate model in dimension 1
model_svm <- diagram_ksvm(diagrams = g,dim = 1,y = y)

# create ten new diagrams with package TDA based on 2D Gaussians
g_new <- lapply(X = 1:10,FUN = function(X){

diag <- TDA::ripsDiag(data.frame(x = rnorm(100,mean = 0,sd = 1),
y = rnorm(100,mean = 0,sd = 1)),
maxscale = 1,
maxdimension = 1)
df <- diagram_to_df(d = diag)
return(df)

})

# predict responses
y_pred <- predict_diagram_ksvm(new_diagrams = g_new,model = model_svm)
}
\seealso{
\code{\link{diagram_ksvm}} for training a SVM model on a training set of persistence diagrams.
}
\author{
Shael Brown - \email{shaelebrown@gmail.com}
}
