% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDAInference.R
\docType{package}
\name{TDAInference}
\alias{TDAInference}
\title{Permutation testing for groups of persistence barcodes}
\description{
This package implements a permutation test which can be used to detect differences
in multiple groups of persistence barcodes - the most common output of the most
common tool in Topological Data Analysis (TDA), persistent homology. Additional
functionality is included to account for between-group dependencies if they are
present. Also, a fast version of wasserstein and bottleneck distances is provided.
}
